generator: Mark QNX `c_void` types as opaque to get raw pointers
Long ago in #339 I added "opaque" types to ensure that pointers to it
stay as pointers in FFI rather than get converted to `&mut` borrows
which don't carry the load of an arbitrary pointer value to "something
unknown" (= opaque).  When the QNX types were added in #429 and #760
some time later, they weren't added to this list and turned into borrows
making them impossible to be handled safely in user code.